ADD: chassis methods add

+++ WarBot/Model/AJBotChassis.h
@@ -12,16 +12,16 @@
 @interface AJBotChassis : NSObject
 
 @property (nonatomic, assign) AJChassisType chassisType;
-@property (nonatomic, assign) float orientation;
-@property (nonatomic, assign) float energy;
+@property (nonatomic, assign) int orientation;
+@property (nonatomic, assign) int energy;
 @property (nonatomic, assign) CGPoint position;
 
 + (id) defaultBotChassis;
 
-- (void) moveForward:(float) distance;
-- (void) moveBackward:(float) distance;
-- (void) turn:(float) angle;
-- (void) turnLeft:(float) angle;
-- (void) turnRight:(float) angle;
+- (void) moveForward:(int) distance;
+- (void) moveBackward:(int) distance;
+- (void) turn:(int) angle;
+- (void) turnLeft:(int) angle;
+- (void) turnRight:(int) angle;
 
 @end
